+++ src/mol-plugin-state/helpers/root-structure.ts
@@ -19,7 +19,7 @@ import { assertUnreachable } from '../../mol-util/type-helpers';
 
 const CommonStructureParams = {
     dynamicBonds: PD.Optional(PD.Boolean(false, { description: 'Ensure bonds are recalculated upon model changes. Also enables calculation of inter-unit bonds in water molecules and ions.' })),
-    mergeSameSymmetryUnits: PD.Optional(PD.Boolean(false, { description: 'Merge structure units that share the same symmetry operator into a single unit.' })),
+    mergeBySymmetry: PD.Optional(PD.Boolean(false, { description: 'Merge structure units that share the same symmetry operator into a single unit.' })),
 };
 type CommonStructureProps = PD.ValuesFor<typeof CommonStructureParams>
 
@@ -110,7 +110,7 @@ export namespace RootStructureDefinition {
         return true;
     }
 
-    async function buildAssembly(plugin: PluginContext, ctx: RuntimeContext, model: Model, id?: string, props?: CommonStructureProps & { merge?: boolean }) {
+    async function buildAssembly(plugin: PluginContext, ctx: RuntimeContext, model: Model, id?: string, props?: CommonStructureProps) {
         let asm: Assembly | undefined = void 0;
 
         const symmetry = ModelSymmetry.Provider.get(model);
@@ -137,7 +137,7 @@ export namespace RootStructureDefinition {
 
         id = asm.id;
         let s = await StructureSymmetry.buildAssembly(base, id!).runInContext(ctx);
-        if (props?.mergeSameSymmetryUnits) s = Structure.mergeUnitsByOperator(s);
+        if (props?.mergeBySymmetry) s = Structure.mergeUnitsByOperator(s);
         const objProps = { label: `Assembly ${id}`, description: Structure.elementDescription(s) };
         return new SO.Molecule.Structure(s, objProps);
     }
@@ -145,7 +145,7 @@ export namespace RootStructureDefinition {
     async function buildSymmetry(ctx: RuntimeContext, model: Model, ijkMin: Vec3, ijkMax: Vec3, props?: CommonStructureProps & { merge?: boolean }) {
         const base = Structure.ofModel(model, props);
         let s = await StructureSymmetry.buildSymmetryRange(base, ijkMin, ijkMax).runInContext(ctx);
-        if (props?.mergeSameSymmetryUnits) s = Structure.mergeUnitsByOperator(s);
+        if (props?.mergeBySymmetry) s = Structure.mergeUnitsByOperator(s);
         const objProps = { label: `Symmetry [${ijkMin}] to [${ijkMax}]`, description: Structure.elementDescription(s) };
         return new SO.Molecule.Structure(s, objProps);
     }
@@ -153,15 +153,15 @@ export namespace RootStructureDefinition {
     async function buildSymmetryMates(ctx: RuntimeContext, model: Model, radius: number, props?: CommonStructureProps & { merge?: boolean }) {
         const base = Structure.ofModel(model, props);
         let s = await StructureSymmetry.builderSymmetryMates(base, radius).runInContext(ctx);
-        if (props?.mergeSameSymmetryUnits) s = Structure.mergeUnitsByOperator(s);
+        if (props?.mergeBySymmetry) s = Structure.mergeUnitsByOperator(s);
         const objProps = { label: `Symmetry Mates`, description: Structure.elementDescription(s) };
         return new SO.Molecule.Structure(s, objProps);
     }
 
-    async function buildSymmetryAssembly(ctx: RuntimeContext, model: Model, generators: StructureSymmetry.Generators, symmetry: Symmetry, props?: CommonStructureProps & { merge?: boolean }) {
+    async function buildSymmetryAssembly(ctx: RuntimeContext, model: Model, generators: StructureSymmetry.Generators, symmetry: Symmetry, props?: CommonStructureProps) {
         const base = Structure.ofModel(model, props);
         let s = await StructureSymmetry.buildSymmetryAssembly(base, generators, symmetry).runInContext(ctx);
-        if (props?.mergeSameSymmetryUnits) s = Structure.mergeUnitsByOperator(s);
+        if (props?.mergeBySymmetry) s = Structure.mergeUnitsByOperator(s);
         const objProps = { label: `Symmetry Assembly`, description: Structure.elementDescription(s) };
         return new SO.Molecule.Structure(s, objProps);
     }